requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
python-dotenv
